LedInterface Class Refactoring
Refactoring for the LedInterface class used in ESP32Marauder. Previously, other boards that were not using Neopixel for the LEDs were in Separate classes. This commit negates the need for this and re-implements the other classes for the Flipper Zero, XIAO-ESP32-S3, & M5StickC boards into the base LedInterface class, allowing for cleaner code in the other Classes & main ino file. Also, changed the mode definitions to an enum class, all modes can be set now with LedMode::<desired-mode> instead of the previous <MODE_DEFINITION> method.